requirements.txt
google-generativeai
requests
python-dotenv
fastapi
uvicorn
httpx
python-multipart
rich
GitPython

# Optional: For Imagen 3 (Tier 1) support
google-cloud-aiplatform